pylot/perception/tracking/nanonets_object_tracking/requirements.txt
opencv_python==4.0.0.21
matplotlib==3.0.3
scipy==1.2.1
tensorflow_gpu==1.13.1
torchvision==0.2.2
torch==1.0.1.post2
imgaug==0.2.8
numpy==1.16.2
Pillow==6.1.0
nonechucks==0.3.1
scikit_learn==0.21.2